% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_2015}
\alias{accident_2015}
\title{Fatality Analysis Reporting System for 2015 year}
\description{
Fatality Analysis Reporting System for 2015 year
}
\references{
\url{https://www-fars.nhtsa.dot.gov/Main/index.aspx}
}
\keyword{data}
