% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Summarize accidents by year and state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{US state index (string)}

\item{years}{A year value (number)}
}
\value{
summarized informaion about accidents for year for state
}
\description{
Summarize accidents by year and state
}
\section{Warning}{

Function will skip year if file with data for year does not exists

Function will stop if invalid state provided
}
\examples{
fars_summarize_years("01", 2001)
}

