% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads FARS (Fatality Analysis Reporting System) data from cvs file.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string. Path to cvs data file.}
}
\value{
Loaded FARS data as \code{\link[dplyr]{tbl_df}}
}
\description{
Reads FARS (Fatality Analysis Reporting System) data from cvs file.
}
\section{Warning}{

Function will stop if provided file not found
}
\examples{
\dontrun{fars_read("accident_2014.csv.bz2")}
}

