% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS (Fatality Analysis Reporting System) data for years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years values (numbers)}
}
\value{
readed FARS data as \code{\link[dplyr]{tbl_df}}
}
\description{
Read FARS (Fatality Analysis Reporting System) data for years
}
\section{Warning}{

Function will skip year if file with data for year does not exists
}
\examples{
\dontrun{fars_read_years(c(2013:2015))}
\dontrun{fars_read_years(c(2013,2015))}
}

