% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize accidents from FARS data by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years values (numbers) to summarize}
}
\value{
Summarized informaion about accidents
}
\description{
Summarize accidents from FARS data by year and month
}
\section{Warning}{

Function will skip year if file with data for year does not exists
}
\examples{
fars_summarize_years(c(2001:2011))
fars_summarize_years(c("2004, 2008"))
}

