% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Genarates archive file name based on year value}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number.}
}
\value{
generated archive file as string
}
\description{
Genarates archive file name based on year value
}
\examples{
make_filename("2011")
make_filename(2011)

}

