% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Display accidents state locations by year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{US state index (string)}

\item{year}{A year value (number)}
}
\value{
this function returns a US state map with dots representing the
fatalities in a given year for a FARS file
}
\description{
Display accidents state locations by year
}
\section{Warning}{

Function will skip year if file with data for year does not exists

Function will stop if invalid state provided
}

