% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Summarize accidents by year and state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{US state index (string)}

\item{year}{A year value (number)}
}
\value{
summarized informaion on map about accidents for year for state as
 \code{\link[maps]{map}}.
}
\description{
Summarize accidents by year and state
}
\section{Warning}{

Function will skip year if file with data for year does not exists

Function will stop if invalid state provided
}

